+++ src/OpenMatchDirector/Handlers/AllocationHandler.cs
@@ -1,0 +1,16 @@
+namespace OpenMatchDirector.Handlers;
+
+/*
+public class AllocationHandler : IChainHandler
+{
+    private IChainHandler Successor;
+    
+    public void Process()
+    {
+        
+        // Allocate
+        var host = AgonesHelper.NewFakeHost();
+        //Metrics - AddAllocationSuccess();
+    }
+}
+*/