+++ CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file. See [versionize](https://github.com/versionize/versionize) for commit guidelines.
 
+<a name="0.2.3"></a>
+## 0.2.3 (2024-06-11)
+
+### Bug Fixes
+
+* correct for backend
+
+<a name="0.2.2"></a>
+## 0.2.2 (2024-06-11)
+
+### Bug Fixes
+
+* correct match function addr/port
+
+<a name="0.2.1"></a>
+## 0.2.1 (2024-06-11)
+
+### Bug Fixes
+
+* add error handling
+
+<a name="0.2.0"></a>
+## 0.2.0 (2024-06-10)
+
+### Features
+
+* adding helpers for openmatch
+
 <a name="0.1.2"></a>
 ## 0.1.2 (2024-06-10)